+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.0-next.32](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.31...@esri/calcite-components-react@3.1.0-next.32) (2025-03-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.1.0-next.31](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.30...@esri/calcite-components-react@3.1.0-next.31) (2025-03-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.1.0-next.30](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.29...@esri/calcite-components-react@3.1.0-next.30) (2025-03-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.1.0-next.29](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.28...@esri/calcite-components-react@3.1.0-next.29) (2025-03-21)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.1.0-next.28](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.27...@esri/calcite-components-react@3.1.0-next.28) (2025-03-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.1.0-next.27](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.26...@esri/calcite-components-react@3.1.0-next.27) (2025-03-19)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.1.0-next.26](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.1.0-next.25...@esri/calcite-components-react@3.1.0-next.26) (2025-03-15)
 
 **Note:** Version bump only for package @esri/calcite-components-react